# fix: reduce memory allocation in dispatch